updating contextId for tN GL.
